parasite: add parasite-compat.h & convert parasite.h types
We need serialization between parasite and criu process in
case of compatible dump. For x86_64 (unsigned long) is 8 bytes,
but for parasite inside compatibe 32-bit application it's 4 bytes.
I changed type declarations for several parasite_args to be
size-specific.

But in some parasite arguments there are standard libc types
which also may differ from native to compatible application
(like timers/timespecs/timevals, etc).
I didn't redefine those structures as size-specific, case it
would need to serialize syscall returns in pie. And pie
should do as lesser as possible in my POV.
So, the serialization work will be done in parasite-syscall,
that should properly check if task compatible and convert
returned structures.
This is done in the very next patch.

All casts added in C files are for gcc to convert from
i.e. u64 to (void*) type which it thinks is incompatible
pointer (and it's right but not this time).

Cc: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>
Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>